+++ plugins/tasks/slack/src/main/java/com/walmartlabs/concord/plugins/slack/SlackChannelTask.java
@@ -21,48 +21,46 @@ package com.walmartlabs.concord.plugins.slack;
  */
 
 import com.walmartlabs.concord.sdk.Context;
+import com.walmartlabs.concord.sdk.ContextUtils;
+import com.walmartlabs.concord.sdk.MapUtils;
 import com.walmartlabs.concord.sdk.Task;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Named;
-import java.util.Arrays;
 import java.util.Map;
 
 import static com.walmartlabs.concord.plugins.slack.SlackClient.Response;
-import static com.walmartlabs.concord.plugins.slack.Utils.assertString;
 
 @Named("slackChannel")
 public class SlackChannelTask implements Task {
 
     private static final Logger log = LoggerFactory.getLogger(SlackChannelTask.class);
 
-    public static final String CHANNEL_ID_KEY = "channelId";
-    public static final String CHANNEL_NAME_KEY = "channelName";
-    public static final String ACTION_KEY = "action";
-    public static final String API_TOKEN_KEY = "apiToken";
-    public static final String RESULT_KEY = "slackChannelId";
+    public static final String SLACK_CHANNEL_ID_KEY = "slackChannelId";
 
     @Override
     public void execute(Context ctx) throws Exception {
         Action action = getAction(ctx);
         log.info("Starting '{}' action...", action);
 
+        Map<String, Object> args = Utils.collectAgs(ctx);
+
         switch (action) {
             case CREATE: {
-                createChannel(ctx);
+                createChannel(ctx, args);
                 break;
             }
             case CREATEGROUP: {
-                createGroup(ctx);
+                createGroup(ctx, args);
                 break;
             }
             case ARCHIVE: {
-                archiveChannel(ctx);
+                archiveChannel(ctx, args);
                 break;
             }
             case ARCHIVEGROUP: {
-                archiveGroup(ctx);
+                archiveGroup(ctx, args);
                 break;
             }
             default:
@@ -70,52 +68,48 @@ public class SlackChannelTask implements Task {
         }
     }
 
-    private static void createChannel(Context ctx) throws Exception {
-        String channelName = assertString(ctx, CHANNEL_NAME_KEY);
-        String apiToken = assertString(ctx, API_TOKEN_KEY);
+    private static void createChannel(Context ctx, Map<String, Object> args) throws Exception {
+        String channelName = MapUtils.assertString(args, TaskParams.CHANNEL_NAME.getKey());
 
-        SlackConfiguration cfg = SlackConfiguration.from(ctx, apiToken);
+        SlackConfiguration cfg = SlackConfiguration.from(args);
         try (SlackClient client = new SlackClient(cfg)) {
             Response r = client.createChannel(channelName);
             handleError(ctx, r, channelName);
 
-            String channelId = extractString(r, "channel", "id");
-            ctx.setVariable(RESULT_KEY, channelId);
-            log.info("Slack channel created: {} -> {} (stored as '{}' variable)", channelName, channelId, RESULT_KEY);
+            String channelId = Utils.extractString(r, "channel", "id");
+            ctx.setVariable(SLACK_CHANNEL_ID_KEY, channelId);
+            log.info("Slack channel created: {} -> {} (stored as '{}' variable)", channelName, channelId, SLACK_CHANNEL_ID_KEY);
         }
     }
 
-    private static void createGroup(Context ctx) throws Exception {
-        String channelName = assertString(ctx, CHANNEL_NAME_KEY);
-        String apiToken = assertString(ctx, API_TOKEN_KEY);
+    private static void createGroup(Context ctx, Map<String, Object> args) throws Exception {
+        String channelName = MapUtils.assertString(args, TaskParams.CHANNEL_NAME.getKey());
 
-        SlackConfiguration cfg = SlackConfiguration.from(ctx, apiToken);
+        SlackConfiguration cfg = SlackConfiguration.from(args);
         try (SlackClient client = new SlackClient(cfg)) {
             Response r = client.createGroupChannel(channelName);
             handleError(ctx, r, channelName);
 
-            String channelId = extractString(r, "group", "id");
-            ctx.setVariable(RESULT_KEY, channelId);
-            log.info("Slack group created: {} -> {} (stored as '{}' variable)", channelName, channelId, RESULT_KEY);
+            String channelId = Utils.extractString(r, "group", "id");
+            ctx.setVariable(SLACK_CHANNEL_ID_KEY, channelId);
+            log.info("Slack group created: {} -> {} (stored as '{}' variable)", channelName, channelId, SLACK_CHANNEL_ID_KEY);
         }
     }
 
-    private static void archiveChannel(Context ctx) throws Exception {
-        String channelId = assertString(ctx, CHANNEL_ID_KEY);
-        String apiToken = assertString(ctx, API_TOKEN_KEY);
+    private static void archiveChannel(Context ctx, Map<String, Object> args) throws Exception {
+        String channelId = MapUtils.assertString(args, TaskParams.CHANNEL_ID.getKey());
 
-        SlackConfiguration cfg = SlackConfiguration.from(ctx, apiToken);
+        SlackConfiguration cfg = SlackConfiguration.from(args);
         try (SlackClient client = new SlackClient(cfg)) {
             Response r = client.archiveChannel(channelId);
             handleError(ctx, r, channelId);
         }
     }
 
-    private static void archiveGroup(Context ctx) throws Exception {
-        String channelId = assertString(ctx, CHANNEL_ID_KEY);
-        String apiToken = assertString(ctx, API_TOKEN_KEY);
+    private static void archiveGroup(Context ctx, Map<String, Object> args) throws Exception {
+        String channelId = MapUtils.assertString(args, TaskParams.CHANNEL_ID.getKey());
 
-        SlackConfiguration cfg = SlackConfiguration.from(ctx, apiToken);
+        SlackConfiguration cfg = SlackConfiguration.from(args);
         try (SlackClient client = new SlackClient(cfg)) {
             Response r = client.archiveGroup(channelId);
             handleError(ctx, r, channelId);
@@ -129,46 +123,9 @@ public class SlackChannelTask implements Task {
         }
     }
 
-    @SuppressWarnings("unchecked")
-    private static String extractString(Response r, String... path) {
-        Map<String, Object> m = r.getParams();
-        if (m == null) {
-            return null;
-        }
-
-        int idx = 0;
-        while (true) {
-            String s = path[idx];
-
-            Object v = m.get(s);
-            if (v == null) {
-                return null;
-            }
-
-            if (idx + 1 >= path.length) {
-                if (v instanceof String) {
-                    return (String) v;
-                } else {
-                    throw new IllegalStateException("Expected a string value @ " + Arrays.toString(path) + ", got: " + v);
-                }
-            }
-
-            if (!(v instanceof Map)) {
-                throw new IllegalStateException("Expected a JSON object, got: " + v);
-            }
-            m = (Map<String, Object>) v;
-
-            idx += 1;
-        }
-    }
-
     private static Action getAction(Context ctx) {
-        Object v = ctx.getVariable(ACTION_KEY);
-        if (v instanceof String) {
-            String s = (String) v;
-            return Action.valueOf(s.trim().toUpperCase());
-        }
-        throw new IllegalArgumentException("'" + ACTION_KEY + "' must be a string");
+        String s = ContextUtils.assertString(ctx, TaskParams.ACTION.getKey());
+        return Action.valueOf(s);
     }
 
     public enum Action {